+++ packages/api/src/objectSet/ObjectSet.test.ts
@@ -1467,6 +1467,13 @@ describe("ObjectSet", () => {
 
       // @ts-expect-error
       objectSet.asType({ type: "object", apiName: "NotImplemented" });
+
+      // Interfaces that don't have any implementedBy fields should still accept any interface
+      const objectSet2 = { asType: () => {} } as unknown as $ObjectSet<
+        FooInterfaceApiTest & { __DefinitionMetadata: { implementedBy: [] } }
+      >;
+
+      objectSet2.asType({ type: "interface", apiName: "NotImplemented" });
     });
     it("restricts casting from object type to interface", () => {
       const objectSet = {} as $ObjectSet<EmployeeApiTest>;